refactor(contract): move TEE measurement votes into api/tee_measurements.rs
