+++ crates/contract/src/api.rs
@@ -9,6 +9,7 @@ pub mod keys;
 pub mod lifecycle;
 pub mod node_migration;
 pub mod sign;
+pub mod tee_measurements;
 pub mod tee_verifier;
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg(test)]
